+++ playwright/tests/george-arent-divisions-landing.spec.ts
@@ -20,5 +20,5 @@ test("verify George Arents division's landing page", async ({ page }) => {
   await expect(
     divisionsLandingPage.collectionsGeorgeArentsHeading
   ).toBeVisible();
-  await expect(divisionsLandingPage.georgeArentsPagination).toBeVisible();
+  await expect(divisionsLandingPage.georgeArentsPagination).toBeHidden();
 });