+++ app/items/[uuid]/page.tsx
@@ -8,6 +8,7 @@ import { ItemPage } from "@/src/components/pages/itemPage/itemPage";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import { CollectionsApi } from "@/src/utils/apiClients/apiClients";
+import { extractAllAnchorsFromHTML } from "@/src/utils/metadata/extractAnchorHrefs";
 
 type ItemProps = {
   params: {
@@ -117,10 +118,12 @@ export default async function ItemViewer({ params, searchParams }: ItemProps) {
       adobeAnalyticsPageName={createAdobeAnalyticsPageName("items", item.title)}
       ga4Data={{
         division: breadcrumbData[1]?.text,
-        collection: breadcrumbData[2]?.text ?? "No detail",
-        subcollection: item.subcollectionName ?? "No detail",
+        collection: breadcrumbData[2]?.text,
+        subcollection: item.subcollectionName ?? undefined,
         contentType: item.contentType,
-        resourceType: item.typeOfResource,
+        resourceType: extractAllAnchorsFromHTML(item.typeOfResource)
+          .map((anchor) => anchor.text.toLowerCase())
+          .join(","),
       }}
     >
       <ItemPage