+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Added Playwright test for Name filter for search-results page (DR-3796)
+- Added Playwright test for George Arents Division's landing page (DR-3716)
+- Added Playwright tests for the About page (DR-3762)
+- Updated Divisions endpoints to use CollectionsAPI (DR-3214)
+- Added relevance to the Collections page (DR-3681)
+- Fixed multiple languages not displaying (DR-3764)
+
 ## [0.4.6] 2025-07-15
 ### Updated
 - Updated IIIF image urls to use v3 instead of v2