+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Updated
 - updated middleware file to redirect synthetic collections (DR-3750)
+- updates node version to 22
+
 ## Added
 - added OpenGraph image to Items page (DR-3791)