+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [1.1.0] 2025-08-28
 ## Updated
 - updated middleware file to redirect synthetic collections (DR-3750)
 - updates node version to 22